% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano plot}
\usage{
volcano_plot(
  x,
  qvalue_cutoff = 0.05,
  meth.diff_cutoff = 25,
  hypermeth_col = "red3",
  hypometh_col = "royalblue3",
  common_col = "seashell3"
)
}
\arguments{
\item{x}{a data frame containing CpG location, strand information and statistics}

\item{qvalue_cutoff}{0.05}

\item{meth.diff_cutoff}{25}

\item{hypermeth_col}{"red3"}

\item{hypometh_col}{"royalblue3"}

\item{common_col}{"seashell3"}
}
\value{
ggplot object
}
\description{
Volcano plot
}
